refactor(ci): improve timeout type and add verification test
- Change timeout parameter from int to float for sub-second precision
- Add test to verify timeout parameter is correctly passed to urlopen
- Update default timeout to 30.0 (float)

Agent-Logs-Url: https://github.com/AigoraCorporation/aigora/sessions/fee16309-8c06-47fb-b543-38a6b5b2ca11

+++ tools/changelog_generator/task_collector.py
@@ -31,9 +31,9 @@ class TaskCollector:
     """
 
     _GITHUB_API = "https://api.github.com"
-    _DEFAULT_TIMEOUT = 30
+    _DEFAULT_TIMEOUT = 30.0
 
-    def __init__(self, token: str, owner: str, repo: str, timeout: int = _DEFAULT_TIMEOUT) -> None:
+    def __init__(self, token: str, owner: str, repo: str, timeout: float = _DEFAULT_TIMEOUT) -> None:
         self._token = token
         self._owner = owner
         self._repo = repo
